feat: rename api route to be more specific (#655)
