+++ pages/api/atlases/[atlasId]/entry-sheet-validations/sync.ts
